Move publishBuild to separate file
