+++ node-src/tasks/verify/index.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, it, vi } from 'vitest';
 
 import { exitCodes } from '../../lib/setExitCode';
 import TestLogger from '../../lib/testLogger';
-import { publishBuild, verifyBuild } from './index';
+import { verifyBuild } from './index';
 
 const environment = {
   CHROMATIC_POLL_INTERVAL: 10,
@@ -10,37 +10,6 @@ const environment = {
   STORYBOOK_VERIFY_TIMEOUT: 20,
 };
 const log = new TestLogger();
-const http = { fetch: vi.fn() };
-
-describe('publishBuild', () => {
-  it('updates the build on the index and updates context', async () => {
-    const announcedBuild = { number: 1, status: 'ANNOUNCED', reportToken: 'report-token' };
-    const publishedBuild = { status: 'PUBLISHED' };
-    const client = { runQuery: vi.fn() };
-    client.runQuery.mockReturnValue({ publishBuild: publishedBuild });
-
-    const ctx = {
-      env: environment,
-      log,
-      http,
-      client,
-      announcedBuild,
-      git: {},
-      sourceDir: '/static/',
-      buildLogFile: 'build-storybook.log',
-      options: {},
-      packageJson: {},
-    } as any;
-    await publishBuild(ctx);
-
-    expect(client.runQuery).toHaveBeenCalledWith(
-      expect.stringMatching(/PublishBuildMutation/),
-      { input: { turboSnapStatus: 'UNUSED' } },
-      { headers: { Authorization: `Bearer report-token` }, retries: 3 }
-    );
-    expect(ctx.announcedBuild).toEqual({ ...announcedBuild, ...publishedBuild });
-  });
-});
 
 describe('verifyBuild', () => {
   const defaultContext = {
